+++ libs/smart-ngrx/src/actions/action.service/load-by-ids.class.spec.ts
@@ -3,13 +3,13 @@ import { fakeAsync, tick } from '@angular/core/testing';
 import { Store } from '@ngrx/store';
 import { BehaviorSubject, Observable, of } from 'rxjs';
 
-import { EffectService } from '../../effects/effect-service';
 import { entityRowsRegistry } from '../../mark-and-delete/entity-rows-registry.class';
 import { effectServiceRegistry } from '../../registrations/effect-service-registry.class';
 import { entityDefinitionCache } from '../../registrations/entity-definition-cache.function';
 import { entityRegistry } from '../../registrations/entity-registry.class';
 import { featureRegistry } from '../../registrations/feature-registry.class';
 import { createStore } from '../../tests/functions/create-store.function';
+import { EffectService } from '../../types/effect-service';
 import { PartialArrayDefinition } from '../../types/partial-array-definition.interface';
 import { SmartNgRXRowBase } from '../../types/smart-ngrx-row-base.interface';
 import { actionFactory } from '../action.factory';
