+++ libs/smart-ngrx/src/index.ts
@@ -2,12 +2,12 @@ export * from './actions/action.factory';
 export * from './common/assert.function';
 export * from './common/cast-to.function';
 export * from './common/for-next.function';
-export * from './effects/effect-service';
 export * from './functions/provide-smart-feature-entities.function';
 export * from './functions/provide-smart-ngrx.function';
 export * from './row-proxy/row-proxy-delete.interface';
 export * from './selector/create-smart-selector.function';
 export * from './selector/smart-array.interface';
+export * from './types/effect-service';
 
 // store.function is needed for testing
 export * from './selector/store.function';
