+++ libs/smart-ngrx/src/registrations/effect-service-registry.class.ts
@@ -1,7 +1,7 @@
 import { InjectionToken } from '@angular/core';
 
 import { assert } from '../common/assert.function';
-import { EffectService } from '../effects/effect-service';
+import { EffectService } from '../types/effect-service';
 import { SmartNgRXRowBase } from '../types/smart-ngrx-row-base.interface';
 
 class EffectServiceRegistry {
