+++ libs/smart-ngrx/src/types/effect-service.token.ts
@@ -1,6 +1,6 @@
 import { InjectionToken } from '@angular/core';
 
-import { EffectService } from '../effects/effect-service';
+import { EffectService } from './effect-service';
 
 /**
  * Shorthand type for the token used to inject the effect service
